requirements.txt
Flask
flask-cors
tensorflow
numpy
Pillow
scikit-learn
matplotlib 